BH-1131: Add relationship field (#203)
* chore: add acm-icons alias for icons

Instead of:
`import Icon from "../../../../../icons";`

We can now do:
`import Icon from "acm-icons";`

From both the settings and publisher app.

* chore: add Relationship icon

* chore: move extra fields below field options

Field options are less important than required fields such as “model to
reference”, so it makes sense to place
them last. It also looks better.

* feat: add the relationship field

* chore: add the additional relationship fields

* style: set dropdown width

* style: use regular font weight for spans in labels

To allow styling of the “(Optional)” text
as in Mason's Figma mockup.

* fix: add Jest module map entry for acm-icons

To prevent “Cannot find module 'acm-icons'…” test failure
due to the introduction of an an alias for the icons package.

* i18n: add translator comments

* chore: make model reference non-editable

Changing the model could leave existing entries in an unexpected state.
(They would refer to an entry from a different model.)

* style: fade help text for read only fields

* test: enable Codeception's “asserts” module

To allow use of `assertEquals()` in acceptance tests.

* test: add relation field acceptance tests

* chore: enable Asserts Codeception module in Circle

* fix: remove unneeded defaultChecked property

* fix: cover the reference model being deleted

Without the optional chain to check for undefined, a field created
with a reference to a removed model will crash the React app
when it's opened for editing because there will be no model
with the selectedReference.

* BH-1132: Relationship field server-side validation (#205)

* chore: check relationship field is valid

* test: verify relationship field must be valid

* chore: show error if related model was deleted

A warning will appear if the selected model was deleted between the
user adding a relationship field and saving it (for example, if another
user deletes the model).

* chore: remove unused import